ENHANCEMENT dry principles
